// src/models/amlich.model.ts
export class AmlichModel{
  day: string;
  month: string;
  year: string;
  leap: string;
  sday: string;
  smonth: string;
  syear: string;
  canchi: string[];
  tietkhi: string;
  GioHoangDao: string;
  GioDauNgay: string;
  LunaDayName: string;
  LunaMonthName: string;
  LunaYearName: string;
  jd: any;
  DayOfWeek: string;
}